+++ co2mpas/datasync.py
@@ -86,7 +86,7 @@ Miscellaneous:
   -V, --version          Print version of the program, with --verbose
                          list release-date and installation details.
   -v, --verbose          Print more verbosely messages - overridden by --logconf.
-  -q, --quite            Print less verbosely messages (warnings) - overridden by --logconf.
+  -q, --quiet            Print less verbosely messages (warnings) - overridden by --logconf.
   --logconf=<conf-file>  Path to a logging-configuration file, according to:
                          See https://docs.python.org/3/library/logging.config.html#configuration-file-format
                          Uses reads a dict-schema if file ends with '.yaml' or '.yml'.
@@ -665,7 +665,7 @@ def _main(*args):
     opts = docopt.docopt(__doc__, argv=args or sys.argv[1:])
 
     verbose = opts['--verbose']
-    quiet = opts['--quite']
+    quiet = opts['--quiet']
     level = None  # Let `init_logging()` decide.
     if verbose:
         level = logging.DEBUG